move CapitalTransferInformation to node/adapter api module
It is no longer shared between hub and node, so it had to be moved.

As a consequence, the CapitalTransferInformation encryption logic had to
be moved as well, but into the node module: EncryptedData is still part
of the hub/node api, on which the node/adapter api should not have any
dependency.

The OASI hash function should be moved in a separate step.
