+++ data-model/src/main/java/ch/prevo/open/data/api/EmploymentCommencement.java
@@ -18,7 +18,6 @@
  ******************************************************************************/
 package ch.prevo.open.data.api;
 
-import ch.prevo.open.encrypted.model.CapitalTransferInformation;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
