+++ data-model/src/main/java/ch/prevo/open/node/data/provider/dummy/NotificationWriter.java
@@ -18,10 +18,10 @@
  ******************************************************************************/
 package ch.prevo.open.node.data.provider.dummy;
 
+import ch.prevo.open.data.api.CapitalTransferInformation;
+import ch.prevo.open.data.api.EmploymentInfo;
 import ch.prevo.open.data.api.FullCommencementNotification;
 import ch.prevo.open.data.api.FullTerminationNotification;
-import ch.prevo.open.data.api.EmploymentInfo;
-import ch.prevo.open.encrypted.model.CapitalTransferInformation;
 
 import java.io.PrintWriter;
