+++ node/src/main/java/ch/prevo/open/node/services/MatchNotificationService.java
@@ -19,14 +19,13 @@
 package ch.prevo.open.node.services;
 
 import ch.prevo.open.data.api.*;
-import ch.prevo.open.encrypted.model.CapitalTransferInformation;
 import ch.prevo.open.encrypted.model.EncryptedData;
 import ch.prevo.open.encrypted.model.MatchForCommencement;
 import ch.prevo.open.encrypted.model.MatchForTermination;
-import ch.prevo.open.encrypted.services.CapitalTransferInfoEncrypter;
 import ch.prevo.open.encrypted.services.Cryptography;
 import ch.prevo.open.node.config.AdapterServiceConfiguration;
 import ch.prevo.open.node.config.NodeConfigurationService;
+import ch.prevo.open.node.crypto.CapitalTransferInfoEncrypter;
 import ch.prevo.open.node.data.provider.EmploymentCommencementProvider;
 import ch.prevo.open.node.data.provider.EmploymentTerminationProvider;
 import ch.prevo.open.node.data.provider.MatchNotificationListener;
