+++ node/src/test/java/ch/prevo/open/node/api/MatchNotificationControllerTest.java
@@ -18,13 +18,13 @@
  ******************************************************************************/
 package ch.prevo.open.node.api;
 
-import ch.prevo.open.encrypted.model.CapitalTransferInformation;
+import ch.prevo.open.data.api.CapitalTransferInformation;
 import ch.prevo.open.encrypted.model.EncryptedData;
 import ch.prevo.open.encrypted.model.MatchForCommencement;
-import ch.prevo.open.encrypted.services.CapitalTransferInfoEncrypter;
 import ch.prevo.open.encrypted.services.Cryptography;
 import ch.prevo.open.node.NodeApplication;
 import ch.prevo.open.node.config.NodeConfigurationService;
+import ch.prevo.open.node.crypto.CapitalTransferInfoEncrypter;
 import ch.prevo.open.node.data.provider.MockProvider;
 import ch.prevo.open.node.data.provider.MockProviderFactory;
 import ch.prevo.open.node.services.MatchNotificationService;
@@ -48,7 +48,7 @@ import java.security.PrivateKey;
 import java.time.LocalDate;
 import java.util.Collections;
 
-import static ch.prevo.open.encrypted.services.DataEncrypter.ASYMMETRIC_TRANSFORMATION_ALGORITHM;
+import static ch.prevo.open.node.crypto.DataEncrypter.ASYMMETRIC_TRANSFORMATION_ALGORITHM;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
