+++ node/src/test/java/ch/prevo/open/node/data/provider/MockProvider.java
@@ -19,7 +19,6 @@
 package ch.prevo.open.node.data.provider;
 
 import ch.prevo.open.data.api.*;
-import ch.prevo.open.encrypted.model.CapitalTransferInformation;
 
 import java.time.LocalDate;
 import java.util.Arrays;
